+++ sample/src/screens/ReduxScreen.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {View, StyleSheet} from 'react-native';
+import {View, StyleSheet, SafeAreaView} from 'react-native';
 
 import Counter from '../components/Counter';